+++ src/widget/index.js
@@ -1,27 +1,24 @@
 import events from 'events';
 
 import Keyless from '@inbloxme/keyless-transactions';
-const KeylessWidget = new Keyless.Widget();
 
 const TokenSwapSDK = require('..');
 
 import { ConnectToWalletModal } from './pages';
 
 import {
-  whetherUserLoggedIn,
+  isUserLoggedIn,
   generateModal,
   getUserPublicAddress,
   updateAuthModal,
   EVENTS
 } from './utils';
 
-export const eventEmitter = new events.EventEmitter();
-
 export class Widget {
-  constructor() {
+  constructor({ rpcURL, env }) {
     const userAddress = getUserPublicAddress();
 
-    this.userLoggedIn = whetherUserLoggedIn();
+    this.userLoggedIn = isUserLoggedIn();
     this.swapVia = '';
     this.handleName = '';
     this.userAddress = userAddress != null ? userAddress : '';
@@ -32,16 +29,18 @@ export class Widget {
     this.activeTab = ConnectToWalletModal();
     this.swapValues = {};
     this.response = {};
-    this.tokenSwap = new TokenSwapSDK.TokenSwap(
-      `wss://ropsten.infura.io/ws/v3/7484a12fa3b544f79bf51ef44edd6db5`
-    );
-    this.keylessWidget = KeylessWidget;
+    this.tokenSwap = new TokenSwapSDK.TokenSwap(rpcURL);
+    this.keylessWidget = new Keyless.Widget({
+      rpcURL,
+      env
+    });
 
     this.isInitialised = false;
     this.EVENTS = EVENTS;
     this.ALL_EVENTS = '*';
     this.ERROR = 'TOKEN_SWAP_ERROR';
     this.INITIALISED_EVENTS = [];
+    this.eventEmitter = new events.EventEmitter();
   }
 
   async initTokenSwap() {
@@ -61,18 +60,18 @@ Widget.prototype.on = function (type, cb) {
 
       if (!this.INITIALISED_EVENTS.includes(evName)) {
         this.INITIALISED_EVENTS.push(evName);
-        eventEmitter.on(evName, cb);
+        this.eventEmitter.on(evName, cb);
       }
     }
   }
 
   if (EVENTS[type] && !this.INITIALISED_EVENTS.includes(EVENTS[type])) {
     this.INITIALISED_EVENTS.push(EVENTS[type]);
-    eventEmitter.on(type, cb);
+    this.eventEmitter.on(type, cb);
   }
 
   if (type === this.ERROR && !this.INITIALISED_EVENTS.includes(this.ERROR)) {
     this.INITIALISED_EVENTS.push(this.ERROR);
-    eventEmitter.on(this.ERROR, cb);
+    this.eventEmitter.on(this.ERROR, cb);
   }
 };